# specs/001-hierarchical-storage-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Hierarchical Storage API
  version: 1.0.0
  description: >
    Backend API for tracking Spaces, recursive Containers, and Items in a
    physical storage hierarchy.
servers:
  - url: http://localhost:3000
paths:
  /auth/register:
    post:
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered and tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      summary: Log in a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login accepted and tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /auth/refresh:
    post:
      summary: Refresh an access token
      operationId: refreshAuthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          $ref: "#/components/responses/Error"
  /auth/logout:
    post:
      summary: Revoke a refresh session
      operationId: logoutUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Session revoked
        "401":
          $ref: "#/components/responses/Error"
  /auth/me:
    get:
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          $ref: "#/components/responses/Error"
  /subscription/status:
    get:
      summary: Get current subscription status
      operationId: getSubscriptionStatus
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription status returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatusResponse"
        "401":
          $ref: "#/components/responses/Error"
  /subscriptions/revenuecat/webhook:
    post:
      summary: Receive RevenueCat subscription event
      operationId: receiveRevenueCatWebhook
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Shared webhook authorization value configured in env.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Event accepted or replay safely ignored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionEventAcceptedResponse"
        "401":
          $ref: "#/components/responses/Error"
  /health:
    get:
      summary: Check API health
      operationId: getHealth
      responses:
        "200":
          description: API health returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /spaces:
    get:
      summary: List spaces
      operationId: listSpaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Spaces returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      summary: Create a space
      operationId: createSpace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceRequest"
      responses:
        "201":
          description: Space created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /spaces/{spaceId}:
    patch:
      summary: Update a space
      operationId: updateSpace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SpaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeRequest"
      responses:
        "200":
          description: Space updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete a space
      operationId: deleteSpace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SpaceId"
      responses:
        "204":
          description: Space deleted
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /spaces/{spaceId}/tree:
    get:
      summary: Get a full or depth-limited space tree
      operationId: getSpaceTree
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SpaceId"
        - $ref: "#/components/parameters/Depth"
      responses:
        "200":
          description: Space tree returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /containers:
    post:
      summary: Create a container
      operationId: createContainer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContainerRequest"
      responses:
        "201":
          description: Container created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /containers/{containerId}:
    patch:
      summary: Update a container
      operationId: updateContainer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContainerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeRequest"
      responses:
        "200":
          description: Container updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete an empty container
      operationId: deleteContainer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContainerId"
      responses:
        "204":
          description: Container deleted
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /containers/{containerId}/tree:
    get:
      summary: Get a container subtree
      operationId: getContainerTree
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContainerId"
        - $ref: "#/components/parameters/Depth"
      responses:
        "200":
          description: Container subtree returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /containers/{containerId}/items:
    get:
      summary: Get items under a container subtree
      operationId: listContainerSubtreeItems
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContainerId"
      responses:
        "200":
          description: Container subtree items returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /containers/{containerId}/move:
    patch:
      summary: Move a container
      operationId: moveContainer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContainerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveNodeRequest"
      responses:
        "200":
          description: Container moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /items:
    post:
      summary: Create an item
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /items/{itemId}:
    get:
      summary: Get an item
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update an item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete an item
      operationId: deleteItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Item deleted
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /items/{itemId}/move:
    patch:
      summary: Move an item
      operationId: moveItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveNodeRequest"
      responses:
        "200":
          description: Item moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /items/{itemId}/path:
    get:
      summary: Get the full item location path
      operationId: getItemPath
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item path returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemPathResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SpaceId:
      name: spaceId
      in: path
      required: true
      schema:
        type: string
    ContainerId:
      name: containerId
      in: path
      required: true
      schema:
        type: string
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
    Depth:
      name: depth
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 25
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
        profileImageUrl:
          type: string
          format: uri
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          minLength: 32
    AuthUser:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        profileImageUrl:
          type: string
          format: uri
    AuthToken:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          minimum: 1
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthTokenResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AuthToken"
    AuthUserResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AuthUser"
    SubscriptionStatus:
      type: object
      required:
        - entitlements
      properties:
        entitlements:
          type: array
          items:
            type: object
            required:
              - key
              - status
            properties:
              key:
                type: string
              status:
                type: string
                enum:
                  - ACTIVE
                  - INACTIVE
                  - EXPIRED
                  - REVOKED
                  - UNKNOWN
              currentPeriodEndsAt:
                type: string
                format: date-time
    SubscriptionStatusResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/SubscriptionStatus"
    SubscriptionEventAcceptedResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - accepted
          properties:
            accepted:
              type: boolean
    NodeType:
      type: string
      enum:
        - SPACE
        - CONTAINER
        - ITEM
    Node:
      type: object
      required:
        - id
        - type
        - name
        - parentId
        - spaceId
        - images
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/NodeType"
        name:
          type: string
        parentId:
          type:
            - string
            - "null"
        spaceId:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/NodeImage"
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TreeNode:
      allOf:
        - $ref: "#/components/schemas/Node"
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: "#/components/schemas/TreeNode"
    PathSegment:
      type: object
      required:
        - id
        - type
        - name
        - images
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/NodeType"
        name:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/NodeImage"
    NodeImage:
      type: object
      required:
        - id
        - url
        - sortOrder
        - createdAt
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        altText:
          type: string
        sortOrder:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    NodeImageInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
        altText:
          type: string
        sortOrder:
          type: integer
          minimum: 0
    CreateSpaceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        metadata:
          type: object
          additionalProperties: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/NodeImageInput"
    CreateContainerRequest:
      type: object
      required:
        - name
        - parentId
      properties:
        name:
          type: string
        parentId:
          type: string
          description: Space or Container parent identifier.
        metadata:
          type: object
          additionalProperties: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/NodeImageInput"
    CreateItemRequest:
      type: object
      required:
        - name
        - parentId
      properties:
        name:
          type: string
        parentId:
          type: string
          description: Space or Container parent identifier.
        metadata:
          type: object
          additionalProperties: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/NodeImageInput"
    UpdateNodeRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        metadata:
          type: object
          additionalProperties: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/NodeImageInput"
    MoveNodeRequest:
      type: object
      required:
        - parentId
      properties:
        parentId:
          type: string
          description: New Space or Container parent identifier.
    NodeResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Node"
    HealthResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - status
          properties:
            status:
              type: string
              enum:
                - ok
    NodeListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Node"
    SpaceListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Node"
    TreeResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/TreeNode"
    ItemPathResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - itemId
            - path
          properties:
            itemId:
              type: string
            path:
              type: array
              items:
                $ref: "#/components/schemas/PathSegment"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INVALID_MOVE
                - CIRCULAR_REFERENCE
                - NOT_FOUND
                - INVALID_PARENT_TYPE
                - SPACE_MISMATCH
                - VALIDATION_ERROR
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties: true